5commit

+++ src/main/java/ru/hogwarts/school/service/FacultyService.java
@@ -1,28 +1,31 @@
 package ru.hogwarts.school.service;
-
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.hogwarts.school.model.Faculty;
 import ru.hogwarts.school.repository.FacultyRepository;
-
 import java.util.*;
 @Service
-    public class FacultyService {
-        private final FacultyRepository facultyRepository;
+public class FacultyService {
+    private final FacultyRepository facultyRepository;
+
     @Autowired
     public FacultyService(FacultyRepository facultyRepository) {
         this.facultyRepository = facultyRepository;
     }
+
     public List<Faculty> findByNameOrColorIgnoreCase(String filter) {
         return facultyRepository.findByNameOrColorIgnoreCase(filter);
     }
-    public Faculty addFaculty(Long id,String name, String color) {
+
+    public Faculty addFaculty(Long id, String name, String color) {
         Faculty faculty = new Faculty(id, name, color);
         return facultyRepository.save(faculty);
     }
+
     public Faculty getFaculty(Long id) {
-        Optional<Faculty> faculty = facultyRepository.findById(id);
-        return faculty.orElse(null);
+        return facultyRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Faculty not found with id: " + id));
     }
     public Faculty updateFaculty(Long id, String name, String color) {
         Optional<Faculty> facultyOpt = facultyRepository.findById(id);
@@ -34,11 +37,14 @@ import java.util.*;
         }
         return null;
     }
-    public void deleteFaculty(Long id) {
-        facultyRepository.deleteById(id);
+    public boolean deleteFaculty(Long id) {
+        if (facultyRepository.existsById(id)) {
+            facultyRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
     public List<Faculty> getAllFaculties() {
         return facultyRepository.findAll();
     }
-
 }
